src/agents/agent_4_dispatch/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
asyncpg>=0.29.0
redis>=5.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0